refactor(frontend): migrate messages/channels tabs to routes (#3962 5.4 PR7)
Extracts the messaging state machinery (messages/channelMessages state,
optimistic pendingMessages merge, channel/direct pagination, and the 8
activeTab-gated scroll/auto-load/mark-as-read effects) out of App.tsx into
a new useMessagingView hook, deletes the now-redundant messages/
channelMessages (+ pagination) state from DataContext, and moves the
channels/messages tabs from activeTab-gated JSX blocks to <Route>
elements — the last pair in the App.tsx tab-to-router migration.

The send/resend/tapback handlers stay in App.tsx (they're entangled with
node lookups, sound/toast/homoglyph settings, and useSourceView's shared
orchestration) but now consume the hook's messages/setMessages/
channelMessages/setChannelMessages/pendingMessagesRef/container refs
under the same names, so their ~700 lines of bodies were unaffected.

App.tsx: 4560 -> 3968 lines. eslint-baseline.json shrinks accordingly
(exhaustive-deps 16->10, no-explicit-any 1->0 for App.tsx).
