+++ src/App.tsx
@@ -54,10 +54,9 @@ import { logger } from './utils/logger';
 // visibleNodeNums/centerMapOnNode to useSourceView (#3962 5.4 PR4)
 import { settingsToMatrix } from './utils/autoAckMatrix';
 import { applyHomoglyphOptimization } from './utils/homoglyph';
-import { playSound, playChannelSound, DEFAULT_SOUND_ID } from './utils/notificationSounds';
 import Sidebar from './components/Sidebar';
 import { SearchModal } from './components/SearchModal/SearchModal.js';
-import { SettingsProvider, useSettings, useNotificationMuteSettings } from './contexts/SettingsContext';
+import { SettingsProvider, useSettings } from './contexts/SettingsContext';
 import { MapProvider, useMapContext } from './contexts/MapContext';
 import type { PositionHistoryItem } from './contexts/MapContext';
 import { DataProvider, useData } from './contexts/DataContext';
@@ -74,6 +73,7 @@ import { useTxStatus } from './hooks/useTxStatus';
 import { useVersionCheck } from './hooks/useVersionCheck';
 import { usePoll, type PollData } from './hooks/usePoll';
 import { useSourceView } from './hooks/useSourceView';
+import { useMessagingView } from './hooks/useMessagingView';
 import { useNotificationNavigationHandler } from './hooks/useNotificationNavigationHandler';
 import LoginModal from './components/LoginModal';
 import LoginPage from './components/LoginPage';
@@ -217,21 +217,14 @@ function App() {
   const showRebootModalRef = useRef<boolean>(false); // Track reboot modal state for interval closure
   const connectionStatusRef = useRef<string>('disconnected'); // Track connection status for interval closure
   const localNodeIdRef = useRef<string>(''); // Track local node ID for immediate access (bypasses React state delay)
-  const pendingMessagesRef = useRef<Map<string, MeshMessage>>(new Map()); // Track pending messages for interval access (bypasses closure stale state)
+  // pendingMessagesRef, the container refs, and channelMessagesRef/messagesRef
+  // moved into useMessagingView (#3962 5.4 PR7) — that hook now owns them and
+  // returns pendingMessagesRef/channelMessagesContainerRef/dmMessagesContainerRef
+  // for the send/resend/tapback handlers below to keep using.
   const homoglyphEnabledRef = useRef<boolean>(false); // Track homoglyph setting for send handlers
 
   // Constants for emoji tapbacks
   const EMOJI_FLAG = 1; // Protobuf flag indicating this is a tapback/reaction
-
-  const channelMessagesContainerRef = useRef<HTMLDivElement>(null);
-  const dmMessagesContainerRef = useRef<HTMLDivElement>(null);
-  const lastScrollLoadTimeRef = useRef<number>(0); // Throttle scroll-triggered loads (200ms)
-  // Track latest channelMessages/messages via refs so the infinite-scroll loaders
-  // see current state without rebuilding their closure on every poll. Without
-  // this the first scroll-up uses a stale offset (0) captured at mount and the
-  // resulting fetch is filtered out by dedup, so the user sees nothing load.
-  const channelMessagesRef = useRef<{ [key: number]: MeshMessage[] }>({});
-  const messagesRef = useRef<MeshMessage[]>([]);
 
   // Detect base URL from pathname
   const detectBaseUrl = () => {
@@ -301,7 +294,6 @@ function App() {
     solarMonitoringLongitude,
     solarMonitoringAzimuth,
     solarMonitoringDeclination,
-    enableAudioNotifications,
     tapbackEmojis,
     setMaxNodeAgeHours,
     setInactiveNodeThresholdHours,
@@ -329,7 +321,9 @@ function App() {
     overlayColors: schemeColors,
   } = useSettings();
 
-  const { isChannelMuted, isDMMuted } = useNotificationMuteSettings();
+  // isChannelMuted/isDMMuted moved into useMessagingView (#3962 5.4 PR7) —
+  // its only consumer (applyPollMessages' notification-sound logic) lives
+  // there now, which calls useNotificationMuteSettings() itself.
 
   // Map context
   // showPaths/showRoute/showMqttNodes/showUdpNodes/showRfNodes/
@@ -359,10 +353,6 @@ function App() {
     setChannels,
     connectionStatus,
     setConnectionStatus,
-    messages,
-    setMessages,
-    channelMessages,
-    setChannelMessages,
     deviceInfo,
     setDeviceInfo,
     deviceConfig,
@@ -371,14 +361,6 @@ function App() {
     setCurrentNodeId,
     nodeAddress,
     setNodeAddress,
-    channelHasMore,
-    setChannelHasMore,
-    channelLoadingMore,
-    setChannelLoadingMore,
-    dmHasMore,
-    setDmHasMore,
-    dmLoadingMore,
-    setDmLoadingMore,
   } = useData();
 
   // Telemetry availability Sets (nodesWithTelemetry/nodesWithWeatherTelemetry/
@@ -386,6 +368,12 @@ function App() {
   // the poll cache (#3962 5.4 PR2) but were consumed ONLY by processedNodes/
   // visibleNodeNums, both now owned by useSourceView (#3962 5.4 PR4), which
   // calls useTelemetryNodes() itself.
+
+  // messages/channelMessages (+ their channelHasMore/channelLoadingMore/
+  // dmHasMore/dmLoadingMore pagination state) were sourced here directly from
+  // DataContext but are not pure poll-cache mirrors — the optimistic-send
+  // merge and infinite-scroll pagination logic they need now live in
+  // useMessagingView (#3962 5.4 PR7), which owns this state itself.
 
   // Consolidated polling for nodes, messages, channels, config
   // Enabled only when connected and not in reboot/user-disconnected state
@@ -518,15 +506,10 @@ function App() {
     unreadCountsData,
   } = useMessaging();
 
-  // The poll callback (processPollData) is memoized without unreadCountsData in
-  // its deps, so bridge the latest filtered counts through a ref. This lets the
-  // per-channel unread badges honor the "Show MQTT/Bridge Messages" toggle the
-  // same way the sidebar dot does (#3787) — the dedicated unread query already
-  // applies excludeMqtt, but the /poll aggregate does not.
-  const unreadCountsDataRef = useRef(unreadCountsData);
-  useEffect(() => {
-    unreadCountsDataRef.current = unreadCountsData;
-  }, [unreadCountsData]);
+  // The unreadCountsData ref-bridge for the poll callback (App's
+  // processPollData is memoized without unreadCountsData in its deps) moved
+  // into useMessagingView (#3962 5.4 PR7) — applyPollMessages needs it, the
+  // favicon effect below reads unreadCountsData directly from context.
 
   // UI context
   const {
@@ -726,8 +709,8 @@ function App() {
   // Track previous total unread count to detect when new messages arrive
   const previousUnreadTotal = useRef<number>(0);
 
-  // Track the newest message ID to detect NEW messages (count-based tracking fails at the 100 message limit)
-  const newestMessageId = useRef<string>('');
+  // newestMessageId moved into useMessagingView (#3962 5.4 PR7) — its only
+  // consumer (applyPollMessages' new-message-sound detection) lives there now.
 
   // Position exchange loading state (separate from traceroute loading)
   const [positionLoading, setPositionLoading] = useState<string | null>(null);
@@ -741,27 +724,8 @@ function App() {
   // Telemetry request loading state
   const [telemetryRequestLoading, setTelemetryRequestLoading] = useState<string | null>(null);
 
-  // Play the notification sound configured for a given channel using the
-  // synthesized Web Audio sound library. Channel `-1` is the DM pseudo-channel;
-  // when no channel is supplied (or it is undefined) the default sound plays,
-  // preserving the previous single-tone behavior. Honors the master audio
-  // toggle exactly as before.
-  const playNotificationSound = useCallback((channelId?: number) => {
-    // Check if audio notifications are enabled
-    if (!enableAudioNotifications) {
-      logger.debug('🔇 Audio notifications disabled, skipping sound');
-      return;
-    }
-
-    logger.debug('🔊 playNotificationSound called for channel:', channelId);
-    if (channelId === undefined) {
-      playSound(DEFAULT_SOUND_ID);
-    } else {
-      // Scope the lookup to the active source so per-source selections (and the
-      // DM pseudo-channel) don't leak across sources that share channel numbers.
-      playChannelSound(channelId, sourceId);
-    }
-  }, [enableAudioNotifications, sourceId]);
+  // playNotificationSound moved into useMessagingView (#3962 5.4 PR7) — its
+  // only consumer (applyPollMessages) lives there now.
 
   // Update favicon with red dot when there are unread messages
   const updateFavicon = useCallback(
@@ -924,6 +888,25 @@ function App() {
     setSelectedRouteSegment,
     setShowPurgeDataModal,
   });
+
+  // Messaging state machinery — messages/channelMessages state (moved off
+  // DataContext), optimistic pendingMessages merge, channel/direct
+  // pagination, and the 8 activeTab-gated scroll/auto-load/mark-as-read
+  // effects (#3962 5.4 PR7, task54_spec.md §3 row 7). The send/resend/
+  // tapback handlers below still consume messages/setMessages/
+  // channelMessages/setChannelMessages/pendingMessagesRef/the two container
+  // refs exactly as they consumed the old DataContext fields — see
+  // src/hooks/useMessagingView.ts.
+  const {
+    messages,
+    setMessages,
+    channelMessages,
+    setChannelMessages,
+    pendingMessagesRef,
+    channelMessagesContainerRef,
+    dmMessagesContainerRef,
+    applyPollMessages,
+  } = useMessagingView();
 
   // Function to detect MQTT/bridge messages that should be filtered
   const isMqttBridgeMessage = (msg: MeshMessage): boolean => {
@@ -1405,392 +1388,10 @@ function App() {
     localStorage.setItem('nodeFilters', JSON.stringify(nodeFilters));
   }, [nodeFilters]);
 
-  // Check if container is scrolled near bottom (within 100px)
-  const isScrolledNearBottom = useCallback((container: HTMLDivElement | null): boolean => {
-    if (!container) return true;
-    const threshold = 100;
-    const { scrollTop, scrollHeight, clientHeight } = container;
-    return scrollHeight - scrollTop - clientHeight < threshold;
-  }, []);
-
-  // Check if container is scrolled near top (within 100px)
-  const isScrolledNearTop = useCallback((container: HTMLDivElement | null): boolean => {
-    if (!container) return false;
-    return container.scrollTop < 100;
-  }, []);
-
-  // Keep refs in sync with the latest state so infinite-scroll loaders read
-  // current values instead of a stale closure. See comment at the ref decls.
-  useEffect(() => {
-    channelMessagesRef.current = channelMessages;
-  }, [channelMessages]);
-
-  useEffect(() => {
-    messagesRef.current = messages;
-  }, [messages]);
-
-  // Load more channel messages (for infinite scroll)
-  const loadMoreChannelMessages = useCallback(async () => {
-    if (channelLoadingMore[selectedChannel] || channelHasMore[selectedChannel] === false) {
-      return;
-    }
-
-    const currentMessages = channelMessagesRef.current[selectedChannel] || [];
-    const offset = currentMessages.length;
-    const container = channelMessagesContainerRef.current;
-
-    // Store scroll position before loading
-    const scrollHeightBefore = container?.scrollHeight || 0;
-
-    setChannelLoadingMore(prev => ({ ...prev, [selectedChannel]: true }));
-
-    try {
-      const result = await api.getChannelMessages(selectedChannel, 100, offset, sourceId);
-
-      if (result.messages.length > 0) {
-        // Process timestamps for new messages
-        const processedMessages = result.messages.map(msg => ({
-          ...msg,
-          timestamp: new Date(msg.timestamp),
-          receivedAt: new Date(msg.receivedAt ?? msg.timestamp),
-        }));
-
-        // Prepend older messages to the existing list, deduplicating by id
-        setChannelMessages(prev => {
-          const existingMessages = prev[selectedChannel] || [];
-          const existingIds = new Set(existingMessages.map(m => m.id));
-          const newMessages = processedMessages.filter(m => !existingIds.has(m.id));
-          return {
-            ...prev,
-            [selectedChannel]: [...newMessages, ...existingMessages],
-          };
-        });
-
-        // Restore scroll position after messages are prepended
-        requestAnimationFrame(() => {
-          if (container) {
-            const scrollHeightAfter = container.scrollHeight;
-            container.scrollTop = scrollHeightAfter - scrollHeightBefore;
-          }
-        });
-      }
-
-      setChannelHasMore(prev => ({ ...prev, [selectedChannel]: result.hasMore }));
-    } catch (error) {
-      logger.error('Failed to load more channel messages:', error);
-      showToast(t('toast.failed_load_older_messages'), 'error');
-    } finally {
-      setChannelLoadingMore(prev => ({ ...prev, [selectedChannel]: false }));
-    }
-  }, [
-    selectedChannel,
-    channelLoadingMore,
-    channelHasMore,
-    setChannelMessages,
-    setChannelHasMore,
-    setChannelLoadingMore,
-    showToast,
-  ]);
-
-  // Load more direct messages (for infinite scroll)
-  const loadMoreDirectMessages = useCallback(async () => {
-    if (!selectedDMNode || !currentNodeId) return;
-
-    const dmKey = [currentNodeId, selectedDMNode].sort().join('_');
-    if (dmLoadingMore[dmKey] || dmHasMore[dmKey] === false) {
-      return;
-    }
-
-    // Get current DM messages from the messages array (channel -1 or direct messages)
-    const currentDMs = messagesRef.current.filter(
-      msg =>
-        (msg.fromNodeId === currentNodeId && msg.toNodeId === selectedDMNode) ||
-        (msg.fromNodeId === selectedDMNode && msg.toNodeId === currentNodeId)
-    );
-    const offset = currentDMs.length;
-    const container = dmMessagesContainerRef.current;
-
-    // Store scroll position before loading
-    const scrollHeightBefore = container?.scrollHeight || 0;
-
-    setDmLoadingMore(prev => ({ ...prev, [dmKey]: true }));
-
-    try {
-      const result = await api.getDirectMessages(currentNodeId, selectedDMNode, 100, offset, sourceId);
-
-      if (result.messages.length > 0) {
-        // Process timestamps for new messages
-        const processedMessages = result.messages.map(msg => ({
-          ...msg,
-          timestamp: new Date(msg.timestamp),
-          receivedAt: new Date(msg.receivedAt ?? msg.timestamp),
-        }));
-
-        // Prepend older messages to the existing list
-        setMessages(prev => {
-          // Remove duplicates by id
-          const existingIds = new Set(prev.map(m => m.id));
-          const newMessages = processedMessages.filter(m => !existingIds.has(m.id));
-          return [...newMessages, ...prev];
-        });
-
-        // Restore scroll position after messages are prepended
-        requestAnimationFrame(() => {
-          if (container) {
-            const scrollHeightAfter = container.scrollHeight;
-            container.scrollTop = scrollHeightAfter - scrollHeightBefore;
-          }
-        });
-      }
-
-      setDmHasMore(prev => ({ ...prev, [dmKey]: result.hasMore }));
-    } catch (error) {
-      logger.error('Failed to load more direct messages:', error);
-      showToast(t('toast.failed_load_older_messages'), 'error');
-    } finally {
-      setDmLoadingMore(prev => ({ ...prev, [dmKey]: false }));
-    }
-  }, [
-    selectedDMNode,
-    currentNodeId,
-    dmLoadingMore,
-    dmHasMore,
-    setMessages,
-    setDmHasMore,
-    setDmLoadingMore,
-    showToast,
-  ]);
-
-  // Handle scroll events to track scroll position (throttled for load-more)
-  const handleChannelScroll = useCallback(() => {
-    if (channelMessagesContainerRef.current) {
-      const atBottom = isScrolledNearBottom(channelMessagesContainerRef.current);
-      setIsChannelScrolledToBottom(atBottom);
-
-      // Check if scrolled near top and trigger load more (throttled to 200ms)
-      const now = Date.now();
-      if (isScrolledNearTop(channelMessagesContainerRef.current) && now - lastScrollLoadTimeRef.current > 200) {
-        lastScrollLoadTimeRef.current = now;
-        void loadMoreChannelMessages();
-      }
-    }
-  }, [isScrolledNearBottom, isScrolledNearTop, loadMoreChannelMessages]);
-
-  const handleDMScroll = useCallback(() => {
-    if (dmMessagesContainerRef.current) {
-      const atBottom = isScrolledNearBottom(dmMessagesContainerRef.current);
-      setIsDMScrolledToBottom(atBottom);
-
-      // Check if scrolled near top and trigger load more (throttled to 200ms)
-      const now = Date.now();
-      if (isScrolledNearTop(dmMessagesContainerRef.current) && now - lastScrollLoadTimeRef.current > 200) {
-        lastScrollLoadTimeRef.current = now;
-        void loadMoreDirectMessages();
-      }
-    }
-  }, [isScrolledNearBottom, isScrolledNearTop, loadMoreDirectMessages]);
-
-  // Attach scroll event listeners. The handler closures are stable now (they
-  // read state via refs), so we re-run this effect when navigation could have
-  // swapped the underlying DOM node — channel switch, DM switch, or tab
-  // switch — to be sure we attach to the freshly mounted container. Without
-  // this the very first scroll on the initial channel never fires the
-  // load-more handler.
-  useEffect(() => {
-    const channelContainer = channelMessagesContainerRef.current;
-    const dmContainer = dmMessagesContainerRef.current;
-
-    if (channelContainer) {
-      channelContainer.addEventListener('scroll', handleChannelScroll);
-    }
-    if (dmContainer) {
-      dmContainer.addEventListener('scroll', handleDMScroll);
-    }
-
-    return () => {
-      if (channelContainer) {
-        channelContainer.removeEventListener('scroll', handleChannelScroll);
-      }
-      if (dmContainer) {
-        dmContainer.removeEventListener('scroll', handleDMScroll);
-      }
-    };
-  }, [handleChannelScroll, handleDMScroll, activeTab, selectedChannel, selectedDMNode]);
-
-  // Force scroll to bottom when channel changes OR when switching to channels tab
-  // Note: We track initial scroll per channel to avoid re-scrolling when user manually scrolls
-  useEffect(() => {
-    if (activeTab === 'channels' && selectedChannel >= 0) {
-      const currentChannelMessages = channelMessages[selectedChannel] || [];
-      const hasMessages = currentChannelMessages.length > 0;
-
-      // Always scroll to bottom when entering the channels tab or changing channels
-      if (hasMessages) {
-        // Use setTimeout to ensure messages are rendered before scrolling
-        setTimeout(() => {
-          if (channelMessagesContainerRef.current) {
-            channelMessagesContainerRef.current.scrollTop = channelMessagesContainerRef.current.scrollHeight;
-            setIsChannelScrolledToBottom(true);
-          }
-        }, 100);
-      }
-    }
-  }, [selectedChannel, activeTab]);
-
-  // Auto-scroll to bottom when new messages arrive and user is already at the bottom
-  const prevChannelMsgCountRef = useRef<Record<number, number>>({});
-  useEffect(() => {
-    const currentMessages = channelMessages[selectedChannel] || [];
-    const prevCount = prevChannelMsgCountRef.current[selectedChannel] || 0;
-    const currentCount = currentMessages.length;
-
-    if (currentCount > prevCount && prevCount > 0) {
-      // New messages arrived — auto-scroll if user was near the bottom
-      const container = channelMessagesContainerRef.current;
-      if (container) {
-        const isNearBottom = container.scrollHeight - container.scrollTop - container.clientHeight < 150;
-        if (isNearBottom) {
-          setTimeout(() => {
-            if (channelMessagesContainerRef.current) {
-              channelMessagesContainerRef.current.scrollTo({
-                top: channelMessagesContainerRef.current.scrollHeight,
-                behavior: 'smooth'
-              });
-            }
-          }, 50);
-        }
-      }
-    }
-
-    prevChannelMsgCountRef.current = {
-      ...prevChannelMsgCountRef.current,
-      [selectedChannel]: currentCount
-    };
-  }, [channelMessages, selectedChannel]);
-
-  // Auto-scroll DMs to bottom when new messages arrive and user is at the bottom
-  const prevDMMsgCountRef = useRef(0);
-  useEffect(() => {
-    const currentCount = messages.length;
-    const prevCount = prevDMMsgCountRef.current;
-
-    if (currentCount > prevCount && prevCount > 0 && activeTab === 'messages') {
-      const container = dmMessagesContainerRef.current;
-      if (container) {
-        const isNearBottom = container.scrollHeight - container.scrollTop - container.clientHeight < 150;
-        if (isNearBottom) {
-          setTimeout(() => {
-            if (dmMessagesContainerRef.current) {
-              dmMessagesContainerRef.current.scrollTo({
-                top: dmMessagesContainerRef.current.scrollHeight,
-                behavior: 'smooth'
-              });
-            }
-          }, 50);
-        }
-      }
-    }
-
-    prevDMMsgCountRef.current = currentCount;
-  }, [messages, activeTab]);
-
-  // Auto-load more channel messages if container doesn't have a scrollbar
-  // This fixes the case where a channel has no recent messages and infinite scroll never triggers
-  useEffect(() => {
-    if (activeTab === 'channels' && selectedChannel >= 0) {
-      // Skip if we're already loading or know there are no more messages
-      if (channelLoadingMore[selectedChannel] || channelHasMore[selectedChannel] === false) {
-        return;
-      }
-
-      // Check after a delay to allow the DOM to render
-      const checkTimer = setTimeout(() => {
-        const container = channelMessagesContainerRef.current;
-        if (container) {
-          // If container doesn't have a scrollbar, load more messages
-          const hasScrollbar = container.scrollHeight > container.clientHeight;
-          if (!hasScrollbar) {
-            void loadMoreChannelMessages();
-          }
-        }
-      }, 200);
-
-      return () => clearTimeout(checkTimer);
-    }
-  }, [selectedChannel, activeTab, channelLoadingMore, channelHasMore, loadMoreChannelMessages]);
-
-  // Force scroll to bottom when DM node changes OR when switching to messages tab
-  useEffect(() => {
-    if (activeTab === 'messages' && selectedDMNode && currentNodeId) {
-      const currentDMMessages = messages.filter(
-        msg =>
-          (msg.fromNodeId === currentNodeId && msg.toNodeId === selectedDMNode) ||
-          (msg.fromNodeId === selectedDMNode && msg.toNodeId === currentNodeId)
-      );
-      const hasMessages = currentDMMessages.length > 0;
-
-      // Always scroll to bottom when entering the messages tab or changing conversations
-      if (hasMessages) {
-        setTimeout(() => {
-          if (dmMessagesContainerRef.current) {
-            dmMessagesContainerRef.current.scrollTop = dmMessagesContainerRef.current.scrollHeight;
-            setIsDMScrolledToBottom(true);
-          }
-        }, 150);
-      }
-    }
-  }, [selectedDMNode, activeTab, currentNodeId]);
-
-  // Auto-load more DM messages if container doesn't have a scrollbar
-  // This fixes the case where a conversation has no recent messages and infinite scroll never triggers
-  useEffect(() => {
-    if (activeTab === 'messages' && selectedDMNode && currentNodeId) {
-      const dmKey = [currentNodeId, selectedDMNode].sort().join('_');
-
-      // Skip if we're already loading or know there are no more messages
-      if (dmLoadingMore[dmKey] || dmHasMore[dmKey] === false) {
-        return;
-      }
-
-      // Check after a delay to allow the DOM to render
-      const checkTimer = setTimeout(() => {
-        const container = dmMessagesContainerRef.current;
-        if (container) {
-          // If container doesn't have a scrollbar, load more messages
-          const hasScrollbar = container.scrollHeight > container.clientHeight;
-          if (!hasScrollbar) {
-            void loadMoreDirectMessages();
-          }
-        }
-      }, 200);
-
-      return () => clearTimeout(checkTimer);
-    }
-  }, [selectedDMNode, activeTab, currentNodeId, dmLoadingMore, dmHasMore, loadMoreDirectMessages]);
-
-  // Unread counts polling is now handled by useUnreadCounts hook in MessagingContext
-
-  // Mark messages as read when viewing a channel — also re-fires when new messages arrive
-  // so that incoming messages are immediately marked as read while the user is viewing the channel.
-  // Without the message count dependency, new messages would show as "unread" until the user
-  // clicks away and back (#2316).
-  const currentChannelMsgCount = (channelMessages[selectedChannel] || []).length;
-  useEffect(() => {
-    if (activeTab === 'channels' && selectedChannel >= 0) {
-      void markMessagesAsRead(undefined, selectedChannel);
-    }
-  }, [selectedChannel, activeTab, markMessagesAsRead, currentChannelMsgCount]);
-
-  // Mark messages as read when viewing a DM conversation — also re-fires on new messages
-  // Filter to only the selected conversation so we don't fire on messages from other DMs
-  const currentDMMsgCount = selectedDMNode
-    ? messages.filter(msg => msg.fromNodeId === selectedDMNode || msg.toNodeId === selectedDMNode).length
-    : 0;
-  useEffect(() => {
-    if (activeTab === 'messages' && selectedDMNode) {
-      void markMessagesAsRead(undefined, undefined, selectedDMNode);
-    }
-  }, [selectedDMNode, activeTab, markMessagesAsRead, currentDMMsgCount]);
+  // Message scroll-position tracking / infinite-scroll pagination / auto-load
+  // / mark-as-read effects (isScrolledNearBottom/isScrolledNearTop through
+  // the DM mark-as-read effect) moved into useMessagingView (#3962 5.4 PR7)
+  // — see src/hooks/useMessagingView.ts.
 
   // Handle push notification navigation (click on notification -> navigate to channel/DM and scroll to message)
   useNotificationNavigationHandler(
@@ -1897,22 +1498,8 @@ function App() {
     };
   }, [connectionStatus]);
 
-  // Timer to update message status indicators (timeout detection after 30s)
-  // Only runs when on channels/messages tabs to reduce CPU usage on mobile (#1769)
-  const [, setStatusTick] = useState(0);
-  useEffect(() => {
-    // Only run timer when viewing messaging tabs where status indicators are visible
-    if (activeTab !== 'channels' && activeTab !== 'messages') {
-      return;
-    }
-
-    const interval = setInterval(() => {
-      // Force re-render to update message status indicators
-      setStatusTick(prev => prev + 1);
-    }, 5000); // Update every 5 seconds (reduced from 1s for mobile performance)
-
-    return () => clearInterval(interval);
-  }, [activeTab]);
+  // The message-status-indicator re-render timer (5s tick, gated to
+  // channels/messages tabs) moved into useMessagingView (#3962 5.4 PR7).
 
   const requestFullNodeDatabase = async () => {
     try {
@@ -2307,203 +1894,13 @@ function App() {
         }
       }
 
-      // Process messages data
+      // Process messages data — optimistic-merge + pagination-preserving
+      // logic lives in useMessagingView (#3962 5.4 PR7); this just feeds it
+      // the poll payload + the two App-local values it needs (localNodeId
+      // computed above, and the currently-selected channel so the unread
+      // count for an open channel zeroes instead of incrementing).
       if (data.messages) {
-        const messagesData = data.messages;
-        const processedMessages = messagesData.map((msg: any) => ({
-          ...msg,
-          timestamp: new Date(msg.timestamp),
-          receivedAt: new Date(msg.receivedAt ?? msg.timestamp),
-        }));
-
-        // Play notification sound if new messages arrived from OTHER users
-        if (processedMessages.length > 0) {
-          const currentNewestMessage = processedMessages[0];
-          const currentNewestId = currentNewestMessage.id;
-
-          if (newestMessageId.current && currentNewestId !== newestMessageId.current) {
-            const isFromOther = currentNewestMessage.fromNodeId !== localNodeId;
-            const isTextMessage = currentNewestMessage.portnum === 1;
-
-            if (isFromOther && isTextMessage) {
-              logger.debug('New message arrived from other user:', currentNewestMessage.fromNodeId);
-              const isDM = currentNewestMessage.channel === -1;
-              const muted = isDM
-                ? isDMMuted(currentNewestMessage.fromNodeId)
-                : isChannelMuted(currentNewestMessage.channel);
-              if (!muted) {
-                // Play the sound configured for this channel (channel -1 is the
-                // DM pseudo-channel, which has its own selectable sound too).
-                playNotificationSound(currentNewestMessage.channel);
-              } else {
-                logger.debug('🔇 Notification sound suppressed (muted):', isDM ? `DM from ${currentNewestMessage.fromNodeId}` : `channel ${currentNewestMessage.channel}`);
-              }
-            }
-          }
-
-          newestMessageId.current = currentNewestId;
-        }
-
-        // Check for matching messages to remove from pending
-        const currentPending = pendingMessagesRef.current;
-        const updatedPending = new Map(currentPending);
-        let pendingChanged = false;
-
-        if (currentPending.size > 0) {
-          currentPending.forEach((pendingMsg, tempId) => {
-            const isDM = pendingMsg.channel === -1;
-
-            const matchingMessage = processedMessages.find((msg: MeshMessage) => {
-              if (msg.text !== pendingMsg.text) return false;
-
-              const senderMatches =
-                (localNodeId && msg.from === localNodeId) ||
-                msg.from === pendingMsg.from ||
-                msg.fromNodeId === pendingMsg.fromNodeId;
-
-              if (!senderMatches) return false;
-              if (Math.abs(msg.timestamp.getTime() - pendingMsg.timestamp.getTime()) >= 30000) return false;
-
-              if (isDM) {
-                const matches =
-                  msg.toNodeId === pendingMsg.toNodeId ||
-                  (msg.to === pendingMsg.to && (msg.channel === 0 || msg.channel === -1));
-                return matches;
-              } else {
-                return msg.channel === pendingMsg.channel;
-              }
-            });
-
-            if (matchingMessage) {
-              updatedPending.delete(tempId);
-              pendingChanged = true;
-            }
-          });
-
-          if (pendingChanged) {
-            pendingMessagesRef.current = updatedPending;
-            setPendingMessages(updatedPending);
-          }
-        }
-
-        // Compute merged messages using setMessages callback to access current state
-        // Preserve older DM messages loaded via infinite scroll (similar to channel messages)
-        const pendingIds = new Set(Array.from(pendingMessagesRef.current.keys()));
-        const pollMsgIds = new Set(processedMessages.map((m: MeshMessage) => m.id));
-
-        setMessages(currentMessages => {
-          // Keep older messages that aren't in the poll (they were loaded via infinite scroll)
-          // Poll returns newest messages, so any messages not in poll are older
-          const olderMsgs = (currentMessages || []).filter(m => {
-            // If message is in poll results, don't keep it (poll version is authoritative)
-            if (pollMsgIds.has(m.id)) return false;
-
-            // For pending messages (temp IDs), only keep if still pending
-            if (m.id.toString().startsWith('temp_')) {
-              if (!pendingIds.has(m.id)) return false;
-              // Safety net: filter out if a matching server message already exists
-              // This catches edge cases where the ref timing or text/sender matching fails
-              // Must use localNodeId fallback (same as primary dedup) because temp messages
-              // created before first poll may have fromNodeId='me' instead of the real node ID
-              const hasServerMatch = processedMessages.some((pm: MeshMessage) =>
-                pm.text === m.text &&
-                ((localNodeId && pm.from === localNodeId) || pm.fromNodeId === m.fromNodeId || pm.from === m.from) &&
-                Math.abs(pm.timestamp.getTime() - m.timestamp.getTime()) < 30000
-              );
-              if (hasServerMatch) return false;
-              return true;
-            }
-
-            // Keep all other older messages (loaded via infinite scroll)
-            return true;
-          });
-
-          // Combine: older messages + poll messages (poll messages are newer/updated)
-          // Sort by timestamp to maintain order
-          const merged = [...olderMsgs, ...processedMessages];
-          merged.sort((a, b) => a.timestamp.getTime() - b.timestamp.getTime());
-          return merged;
-        });
-
-        // Group messages by channel (use processedMessages since we don't need pending for channel groups)
-        const channelGroups: { [key: number]: MeshMessage[] } = {};
-        processedMessages.forEach((msg: MeshMessage) => {
-          if (msg.channel === -1) return;
-          if (!channelGroups[msg.channel]) {
-            channelGroups[msg.channel] = [];
-          }
-          channelGroups[msg.channel].push(msg);
-        });
-
-        // Update unread counts from the dedicated (filtered) unread query rather
-        // than the raw /poll aggregate, so the per-channel badges respect the
-        // "Show MQTT/Bridge Messages" toggle just like the sidebar dot (#3787).
-        // Fall back to the poll payload only until the first dedicated fetch lands.
-        const currentSelected = selectedChannelRef.current;
-        const newUnreadCounts: { [key: number]: number } = {};
-
-        const filteredChannelUnreads = unreadCountsDataRef.current?.channels ?? data.unreadCounts?.channels;
-        if (filteredChannelUnreads) {
-          Object.entries(filteredChannelUnreads).forEach(([channelId, count]) => {
-            const chId = parseInt(channelId, 10);
-            if (chId === currentSelected) {
-              newUnreadCounts[chId] = 0;
-            } else {
-              newUnreadCounts[chId] = count as number;
-            }
-          });
-        }
-
-        setUnreadCounts(newUnreadCounts);
-
-        // Merge poll messages with existing messages (preserve older messages loaded via infinite scroll)
-        setChannelMessages(prev => {
-          const merged: { [key: number]: MeshMessage[] } = {};
-
-          // Get all channel IDs from both existing and new messages
-          const allChannelIds = new Set([...Object.keys(prev).map(Number), ...Object.keys(channelGroups).map(Number)]);
-
-          allChannelIds.forEach(channelId => {
-            const existingMsgs = prev[channelId] || [];
-            const pollMsgs = channelGroups[channelId] || [];
-
-            // Create a map of poll message IDs for quick lookup
-            const pollMsgIds = new Set(pollMsgs.map(m => m.id));
-
-            // Keep older messages that aren't in the poll (they were loaded via infinite scroll)
-            // Poll returns newest 100, so any messages not in poll are older
-            // Also filter out pending messages that are no longer pending (they've been matched to real messages)
-            const olderMsgs = existingMsgs.filter(m => {
-              // If message is in poll results, don't keep it (poll version is authoritative)
-              if (pollMsgIds.has(m.id)) return false;
-
-              // For pending messages (temp IDs), only keep if still pending
-              // Once matched/acknowledged, pendingIds won't contain it anymore
-              // Channel messages use 'temp_' prefix, DMs use 'temp_dm_' prefix
-              if (m.id.toString().startsWith('temp_')) {
-                if (!pendingIds.has(m.id)) return false;
-                // Safety net: filter out if a matching server message already exists
-                // Must use localNodeId fallback (same as primary dedup) because temp messages
-                // created before first poll may have fromNodeId='me' instead of the real node ID
-                const hasServerMatch = pollMsgs.some(pm =>
-                  pm.text === m.text &&
-                  ((localNodeId && pm.from === localNodeId) || pm.fromNodeId === m.fromNodeId || pm.from === m.from) &&
-                  Math.abs(pm.timestamp.getTime() - m.timestamp.getTime()) < 30000
-                );
-                if (hasServerMatch) return false;
-                return true;
-              }
-
-              // Keep all other older messages (loaded via infinite scroll)
-              return true;
-            });
-
-            // Combine: older messages + poll messages (poll messages are newer/updated)
-            merged[channelId] = [...olderMsgs, ...pollMsgs];
-          });
-
-          return merged;
-        });
+        applyPollMessages(data, localNodeId, selectedChannelRef.current);
       }
 
       // Process config data
@@ -2538,7 +1935,7 @@ function App() {
         setTraceroutes(data.traceroutes);
       }
     },
-    [currentNodeId, playNotificationSound, setTraceroutes, isChannelMuted, isDMMuted]
+    [currentNodeId, setTraceroutes, applyPollMessages]
   );
 
   // Process poll data when it changes (from usePoll hook)
@@ -4078,14 +3475,17 @@ function App() {
         )}
 
         {/*
-          Tab region nested Routes (#3962 5.4 PR1). Each tab is migrating from
-          an `activeTab === 'x' && …` render gate to its own <Route> one PR at
-          a time (task54_spec.md §3). Un-migrated tabs keep rendering via the
-          `path="*"` fallback below, unchanged, driven by the activeTab<->route
-          adapter in UIContext. `audit` was the PR1 proof leaf; PR3 migrated
-          the other leaf tabs; PR4 migrated `nodes` (the default/index tab)
-          alongside extracting its shared node/traceroute/map orchestration
-          into `useSourceView` (src/hooks/useSourceView.ts).
+          Tab region nested Routes (#3962 5.4 PR1). All 15 tabs are now
+          `<Route>` elements (task54_spec.md §3) — the `path="*"` fallback
+          below renders nothing, preserving prior blank-fallback behavior for
+          an unrecognized sub-path. `audit` was the PR1 proof leaf; PR3
+          migrated the other leaf tabs; PR4 migrated `nodes` (the default/
+          index tab) alongside extracting its shared node/traceroute/map
+          orchestration into `useSourceView` (src/hooks/useSourceView.ts);
+          PR5 migrated info/dashboard/configuration; PR6 migrated settings/
+          automation; PR7 migrated channels/messages alongside extracting
+          their messaging state machinery into `useMessagingView`
+          (src/hooks/useMessagingView.ts) — see that file's doc comment.
         */}
         <Routes>
           <Route index element={<Navigate to="nodes" replace />} />
@@ -4299,9 +3699,10 @@ function App() {
               </ErrorBoundary>
             }
           />
-          <Route path="*" element={<>
-        {activeTab === 'channels' && (
-          <ErrorBoundary fallbackTitle="Channels failed to load">
+          <Route
+            path="channels"
+            element={
+              <ErrorBoundary fallbackTitle="Channels failed to load">
           <ChannelsTab
             channels={channels}
             channelDatabaseEntries={channelDatabaseEntries}
@@ -4347,10 +3748,13 @@ function App() {
             onFocusMessageHandled={() => setFocusMessageId(null)}
             mqttReadOnly={isMqttBridge}
           />
-          </ErrorBoundary>
-        )}
-        {activeTab === 'messages' && (
-          <ErrorBoundary fallbackTitle="Messages failed to load">
+              </ErrorBoundary>
+            }
+          />
+          <Route
+            path="messages"
+            element={
+              <ErrorBoundary fallbackTitle="Messages failed to load">
           <MessagesTab
             processedNodes={processedNodes}
             nodes={nodes}
@@ -4425,15 +3829,19 @@ function App() {
               }
             }}
           />
-          </ErrorBoundary>
-        )}
-        {/* 'audit' migrated to <Route path="audit"> above (#3962 5.4 PR1 proof leaf) */}
-        {/* 'notifications', 'users', 'admin', 'security', 'mqtt-config', 'packetmonitor'
-            migrated to <Route> elements above (#3962 5.4 PR3 leaf tab group) */}
-        {/* 'automation', 'settings' migrated to <Route> elements above (#3962 5.4 PR6) */}
-        {/* 'info', 'dashboard', 'configuration' migrated to <Route> elements above
-            (#3962 5.4 PR5) */}
-          </>} />
+              </ErrorBoundary>
+            }
+          />
+          {/* 'audit' migrated to <Route path="audit"> above (#3962 5.4 PR1 proof leaf) */}
+          {/* 'notifications', 'users', 'admin', 'security', 'mqtt-config', 'packetmonitor'
+              migrated to <Route> elements above (#3962 5.4 PR3 leaf tab group) */}
+          {/* 'automation', 'settings' migrated to <Route> elements above (#3962 5.4 PR6) */}
+          {/* 'info', 'dashboard', 'configuration' migrated to <Route> elements above
+              (#3962 5.4 PR5) */}
+          {/* 'channels', 'messages' migrated to <Route> elements above (#3962 5.4 PR7) —
+              every tab is now a route; this catch-all preserves the prior blank-fallback
+              behavior for an unrecognized sub-path. */}
+          <Route path="*" element={null} />
         </Routes>
       </main>
